core: move DB to separate file for import
