+++ app/models.py
@@ -3,15 +3,14 @@ import time
 import uuid
 from datetime import datetime
 
-#from app.factory import db
 from faker import Faker
 from flask import current_app
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import AnonymousUserMixin, UserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
+from app.db import db
 
 fake = Faker()
-db = SQLAlchemy()
 
 class User(UserMixin, db.Model):
